Add dep fetch diagnostics & refactor
Dependency fetching had grown into Compile alongside project path and
Build.act resolution, which made proxy diagnostics harder to own without
expanding the compile pipeline module.

Move project, path, and build-spec helpers into Acton.Project and make
Acton.FetchDependencies the direct owner of package and Zig dependency
fetching.

Fetch failures report proxy environment, selected proxy or direct decisions,
cache state, HTTP failure class, and Zig fallback context.
