+++ compiler/lib/src/Acton/Compile.hs
@@ -82,9 +82,7 @@ TODO:
     (e.g. long-running B) are not discarded when A changes.
 -}
 module Acton.Compile
-  ( Paths(..)
-  , ProjCtx(..)
-  , TaskKey(..)
+  ( TaskKey(..)
   , GlobalTask(..)
   , CompileTask(..)
   , BackInput(..)
@@ -134,9 +132,6 @@ module Acton.Compile
   , backPassFailureMessage
   , BackJobResult(..)
   , defaultCompileOptions
-  , ProjectError(..)
-  , throwProjectError
-  , fetchDependencies
   , parseActSource
   , parseActSnapshot
   , parseActFile
@@ -144,7 +139,6 @@ module Acton.Compile
   , readModuleDoc
   , readModuleDocIndexEntry
   , readImports
-  , special_projects
   , buildGlobalTasks
   , selectNeededTasks
   , selectAffectedTasks
@@ -155,32 +149,7 @@ module Acton.Compile
   , shouldGenerateDocOutput
   , runBackPasses
   , runBackJobs
-  , findProjectDir
-  , BackgroundCompilerLock
-  , backgroundCompilerLockPath
-  , tryBackgroundCompilerLock
-  , releaseBackgroundCompilerLock
-  , projectLockPath
-  , withProjectLockOnWait
-  , withProjectLock
-  , findPaths
-  , discoverProjects
-  , pathsForModule
-  , searchPathForProject
-  , moduleNameFromFile
-  , enumerateProjectModules
-  , normalizeDepOverrides
-  , applyDepOverrides
-  , collectDepTypePaths
-  , depTypePathsFromMap
-  , resolveDepBase
-  , loadBuildSpec
   , filterActFile
-  , srcFile
-  , outBase
-  , tyDbPath
-  , srcBase
-  , getModPath
   , modNameToFilename
   , modNameToString
   , nameToString
@@ -194,10 +163,6 @@ module Acton.Compile
   , getImplHashCached
   , updatePubHashCache
   , rootEligible
-  , normalizePathSafe
-  , isAbsolutePath
-  , collapseDots
-  , rebasePath
   ) where
 
 import Prelude hiding (readFile, writeFile)
@@ -227,25 +192,24 @@ import Acton.Builtin (mBuiltin)
 import Acton.Prim (mPrim)
 import qualified Acton.BuildSpec as BuildSpec
 import qualified Acton.DocPrinter as DocP
-import qualified Acton.Fingerprint as Fingerprint
+import qualified Acton.FetchDependencies as Fetch
 import qualified Acton.Diagnostics as Diag
+import Acton.Project
 import qualified Acton.SourceProvider as Source
 import Utils
 import qualified Pretty
 import Pretty hiding ((<>))
 import qualified InterfaceFiles
 
-import Control.Applicative ((<|>))
 import Control.Concurrent.Async
 import Control.Concurrent.MVar
 import Control.Concurrent (forkIO, myThreadId, threadCapability, threadDelay)
 import Control.Concurrent.STM (TChan, TVar, atomically, check, modifyTVar', newTChanIO, newTVarIO, readTChan, readTVar, writeTChan)
 import Control.DeepSeq (rnf)
-import Control.Exception (Exception, IOException, SomeAsyncException, SomeException, bracketOnError, catch, displayException, evaluate, finally, fromException, mask_, throwIO, try)
+import Control.Exception (Exception, IOException, SomeAsyncException, SomeException, catch, displayException, evaluate, finally, fromException, mask_, throwIO, try)
 import Control.Monad
 import Data.Binary (encode)
-import Data.Bits (shiftL, shiftR, (.|.))
-import Data.Char (isAlpha, isDigit, isHexDigit, isSpace, toLower)
+import Data.Char (isHexDigit, isSpace)
 import Data.Either (partitionEithers)
 import Data.Graph
 import Data.List (find, foldl', intercalate, intersperse, isPrefixOf, isSuffixOf, nub, partition)
@@ -257,27 +221,16 @@ import qualified Data.Map as M
 import Data.Ord (Down(..))
 import qualified Data.Set
 import Data.Time.Clock (UTCTime)
-import Data.Word (Word8, Word32, Word64)
 import Error.Diagnose (Diagnostic)
 import GHC.Conc (getNumCapabilities)
-import qualified Network.HTTP.Client as HTTP
-import Network.HTTP.Client.TLS (tlsManagerSettings)
-import Network.HTTP.Types.Header (hContentDisposition, hContentType)
-import Network.HTTP.Types.Status (statusCode)
 import System.Clock
 import System.Directory
-import System.Directory.Recursive (getFilesRecursive)
-import System.Environment (getExecutablePath, lookupEnv)
-import System.FileLock (FileLock, SharedExclusive(Exclusive), tryLockFile, unlockFile, withFileLock)
+import System.Environment (getExecutablePath)
 import System.FilePath ((</>))
 import System.FilePath.Posix
-import System.Exit (ExitCode(..))
 import System.IO hiding (readFile, writeFile)
-import System.IO.Temp (createTempDirectory, withSystemTempDirectory)
 import System.IO.Unsafe (unsafePerformIO)
 import System.Posix.Files (FileStatus, deviceID, fileID, fileSize, getFileStatus, modificationTimeHiRes, statusChangeTimeHiRes)
-import System.Process (CreateProcess(cwd), readCreateProcessWithExitCode, proc)
-import System.Random (randomRIO)
 import Text.PrettyPrint (renderStyle, style, Style(..), Mode(PageMode))
 import Text.Show.Pretty (ppDoc)
 import Text.Printf
@@ -289,10 +242,6 @@ import qualified Data.ByteString.Base16 as Base16
 import qualified Crypto.Hash.SHA256 as SHA256
 
 
-newtype ProjectError = ProjectError String deriving (Show)
-
-instance Exception ProjectError
-
 newtype DbpSelectionError = DbpSelectionError String deriving (Show)
 
 instance Exception DbpSelectionError where
@@ -302,12 +251,6 @@ newtype FrontOutputError = FrontOutputError String deriving (Show)
 
 instance Exception FrontOutputError where
   displayException (FrontOutputError msg) = msg
-
--- | Raise a ProjectError for library callers.
--- Used by project discovery and path helpers to stop on unrecoverable errors.
-throwProjectError :: String -> IO a
-throwProjectError msg = throwIO (ProjectError msg)
-
 
 data CompileFailure
   = CompileCycleFailure String
@@ -677,7 +620,7 @@ prepareCompilePlan sp gopts sched opts srcFiles allowPrune mChangedPaths = do
   ctx <- prepareCompileContext opts srcFiles
   specChanged <- checkBuildSpecChange sched (ccBuildStamp ctx)
   when specChanged $
-    fetchDependencies gopts (ccPathsRoot ctx) (ccDepOverrides ctx)
+    Fetch.fetchDependencies gopts (ccPathsRoot ctx) (ccDepOverrides ctx)
   let mChanged = if specChanged then Nothing else mChangedPaths
   prepareCompilePlanFromContext sp gopts ctx srcFiles allowPrune mChanged
 
@@ -887,12 +830,6 @@ dump :: A.ModName -> String -> String -> IO ()
 dump mn h txt =
   putStrLn ("\n\n== " ++ h ++ ": " ++ modNameToString mn ++ " ================================\n" ++ txt
             ++ '\n' : replicate (38 + length h + length (modNameToString mn)) '=' ++ "\n")
-
--- | Compute the subdirectory for a module within a types root.
--- Used when creating directories before writing .tydb files.
-getModPath :: FilePath -> A.ModName -> FilePath
-getModPath path mn =
-  joinPath [path, joinPath $ init $ A.modPath mn]
 
 -- Global caches (process‑wide) to reduce repeated .tydb lookups during parallel builds
 --
@@ -1565,8 +1502,6 @@ resolveProviders key proj order modSets deps imps =
         dropLib mn      = case A.modPath mn of
                             [n,"lib"] | Data.Set.member n deps -> A.modName [n]
                             _ -> mn
-
-special_projects = ["", "base", "acton_scratch"]
 
 type ProjDir = FilePath
 type ActFile = FilePath
@@ -4331,1095 +4266,6 @@ runBackJobs gopts maxPar onStart onDone jobs = do
         Just (job, mline) -> do
           onDone job mline
           flushReady (M.delete ix res) (ix + 1)
-
-
--- Paths handling -------------------------------------------------------------------------------------
-
-data Paths      = Paths {
-                    searchPath  :: [FilePath],
-                    sysPath     :: FilePath,
-                    sysTypes    :: FilePath,
-                    projPath    :: FilePath,
-                    projOut     :: FilePath,
-                    projTypes   :: FilePath,
-                    binDir      :: FilePath,
-                    srcDir      :: FilePath,
-                    isTmp       :: Bool,
-                    fileExt     :: String,
-                    modName     :: A.ModName,
-                    projName    :: String
-                  }
-
--- Per-project context used for multi-project orchestration.
--- Identified by projRoot (absolute path). Keeps directories and BuildSpec if present.
-data ProjCtx = ProjCtx {
-                     projRoot    :: FilePath,
-                     projOutDir  :: FilePath,
-                     projTypesDir:: FilePath,
-                     projSrcDir  :: FilePath,
-                     projSysPath :: FilePath,
-                     projSysTypes:: FilePath,
-                     projBuildSpec :: BuildSpec.BuildSpec,
-                     projDeps     :: [(String, FilePath)]          -- resolved dependency roots (abs paths)
-                   } deriving (Show)
-
-systemTypePaths :: FilePath -> FilePath -> [FilePath]
-systemTypePaths sys types =
-    [joinPath [sys, "std", "out", "types"], types]
-
-type FingerprintMap = M.Map String FilePath
-
-scratchBuildSpec :: FilePath -> BuildSpec.BuildSpec
-scratchBuildSpec projRoot =
-    let name = "acton_scratch"
-        prefix = Fingerprint.fingerprintPrefixForName name
-        suffix = Fingerprint.fingerprintPrefixForName projRoot
-        fp = Fingerprint.formatFingerprint (Fingerprint.updateFingerprintPrefix prefix (fromIntegral suffix))
-    in BuildSpec.BuildSpec
-         { BuildSpec.specName = name
-         , BuildSpec.specDescription = Nothing
-         , BuildSpec.fingerprint = fp
-         , BuildSpec.dependencies = M.empty
-         , BuildSpec.zig_dependencies = M.empty
-         , BuildSpec.libraries = M.empty
-         }
-
-
-normalizeFingerprintKey :: String -> Maybe String
-normalizeFingerprintKey raw =
-    Fingerprint.formatFingerprint <$> Fingerprint.parseFingerprint raw
-
-fingerprintKeyFromSpec :: BuildSpec.BuildSpec -> String
-fingerprintKeyFromSpec spec =
-    case normalizeFingerprintKey (BuildSpec.fingerprint spec) of
-      Just fp -> fp
-      Nothing -> BuildSpec.fingerprint spec
-
-applyFingerprint :: FilePath -> BuildSpec.BuildSpec -> FingerprintMap -> (FilePath, FingerprintMap, String)
-applyFingerprint path spec fpMap =
-    let fp = fingerprintKeyFromSpec spec
-    in case M.lookup fp fpMap of
-         Just canonical -> (canonical, fpMap, fp)
-         Nothing -> (path, M.insert fp path fpMap, fp)
-
--- | Discover all projects reachable from a root project.
--- Follows Build.act dependencies, applies overrides/pins, and
--- returns a map from project root to ProjCtx while skipping duplicates.
-discoverProjects :: C.GlobalOptions -> FilePath -> FilePath -> [(String, FilePath)] -> IO (M.Map FilePath ProjCtx)
-discoverProjects gopts sysAbs rootProj depOverrides = do
-    rootAbs <- normalizePathSafe rootProj
-    rootSpec0 <- loadBuildSpec rootAbs
-    rootSpec  <- applyDepOverrides rootAbs depOverrides rootSpec0
-    let rootPins = BuildSpec.dependencies rootSpec
-        (_, fpMap0, _) = applyFingerprint rootAbs rootSpec M.empty
-    fst <$> go rootAbs Data.Set.empty M.empty fpMap0 rootPins rootAbs (Just rootSpec)
-  where
-    go root seen acc fpMap pins dir mSpec = do
-      dirAbs <- normalizePathSafe dir
-      if Data.Set.member dirAbs seen
-        then return (acc, fpMap)
-        else do
-          spec0 <- case mSpec of
-                     Just s -> return s
-                     Nothing -> loadBuildSpec dirAbs
-          spec <- applyDepOverrides dirAbs depOverrides spec0
-          let (_, fpMap1, _) = applyFingerprint dirAbs spec fpMap
-          (deps, fpMap2) <- do
-            let depsList = M.toList (BuildSpec.dependencies spec)
-            foldM (collectDep pins dirAbs) ([], fpMap1) depsList
-          let outDir   = joinPath [dirAbs, "out"]
-              typesDir = joinPath [outDir, "types"]
-              srcDir'  = joinPath [dirAbs, "src"]
-              ctx = ProjCtx { projRoot = dirAbs
-                            , projOutDir = outDir
-                            , projTypesDir = typesDir
-                            , projSrcDir = srcDir'
-                            , projSysPath = sysAbs
-                            , projSysTypes = joinPath [sysAbs, "base", "out", "types"]
-                            , projBuildSpec = spec
-                            , projDeps = [ (n, p) | (n, p, _) <- reverse deps ]
-                            }
-              acc' = M.insert dirAbs ctx acc
-              seen' = Data.Set.insert dirAbs seen
-          foldM (step root seen' pins) (acc', fpMap2) (reverse deps)
-
-    collectDep pins base (accDeps, fpMap) (depName, dep) = do
-      let (chosenDep, conflict) =
-            case M.lookup depName pins of
-              Nothing -> (dep, Nothing)
-              Just pinDep ->
-                if pinDep == dep
-                  then (dep, Nothing)
-                  else (pinDep, Just dep)
-      when (isJust conflict) $
-        unless (C.quiet gopts) $
-          putStrLn ("Warning: dependency '" ++ depName ++ "' in " ++ base
-                    ++ " overridden by root pin")
-      depBase <- resolveDepBase base depName chosenDep
-      depAbs  <- normalizePathSafe depBase
-      (depPath, fpMap', depSpec) <- canonicalizeDep depAbs fpMap
-      return ((depName, depPath, depSpec) : accDeps, fpMap')
-
-    canonicalizeDep depAbs fpMap = do
-      spec0 <- loadBuildSpec depAbs
-      spec <- applyDepOverrides depAbs depOverrides spec0
-      let (canonPath, fpMap', fp) = applyFingerprint depAbs spec fpMap
-      when (canonPath /= depAbs) $
-        unless (C.quiet gopts) $
-          putStrLn ("Warning: dependency fingerprint " ++ fp
-                    ++ " at " ++ depAbs ++ " deduplicated to " ++ canonPath)
-      let depSpec = if canonPath == depAbs then Just spec else Nothing
-      return (canonPath, fpMap', depSpec)
-
-    step root seen pins (acc, fpMap) (_, depBase, mSpec) =
-      go root seen acc fpMap pins depBase mSpec
-
--- Given a FILE and optionally --syspath PATH:
--- 'sysPath' is the path to the system directory as given by PATH, defaulting to the acton executable directory.
--- 'sysTypes' is directory "types" under 'sysPath'.
--- 'projPath' is the closest parent directory of FILE that contains a Build.act
--- file, or a temporary directory in "/tmp" if no such parent exists.
--- 'projOut' is directory "out" under 'projPath'.
--- 'projTypes' is directory "types" under 'projOut'.
--- 'binDir' is the directory prefix of FILE if 'projPath' is temporary, otherwise it is directory "bin" under 'projOut'
--- 'srcDir' is the directory prefix of FILE if 'projPath' is temporary, otherwise it is directory "src" under 'projPath'
--- 'fileExt' is file suffix of FILE.
--- 'modName' is the module name of FILE (its path after 'src' except 'fileExt', split at every '/')
-
--- | Compute the source file path for a module under its project src dir.
-srcFile                 :: Paths -> A.ModName -> FilePath
-srcFile paths mn        = srcBase paths mn ++ ".act"
-
--- | Compute the output base path (without extension) for a module.
--- Used to locate .tydb/.c/.h output under the project's types directory.
-outBase                 :: Paths -> A.ModName -> FilePath
-outBase paths mn        = joinPath (projTypes paths : A.modPath mn)
-
-tyDbPath                :: Paths -> A.ModName -> FilePath
-tyDbPath paths mn       = outBase paths mn ++ InterfaceFiles.interfaceExt
-
--- | Compute the module path without extension under the project's src dir.
--- Used to derive the .act path or related per-module files.
-srcBase                 :: Paths -> A.ModName -> FilePath
-srcBase paths mn        = --trace ("## srcBase " ++ prstr mn ++ " --> " ++ fpath) $
-                          fpath
-  where fpath           = joinPath (srcDir paths : names)
-        names           = case A.modPath mn of
-                            n:ns | n == projName paths -> ns
-                            ns -> ns
-
-
--- | Walk upward from a path to find a project root.
--- A project root is identified by Build.act plus a
--- src/ directory; returns Nothing if we reach filesystem root.
-
--- | Check whether a directory is an Acton project root.
--- Requires Build.act and a src/ directory.
-isActonProjectRoot :: FilePath -> IO Bool
-isActonProjectRoot path = do
-    runacton <- lookupEnv "ACTON_RUNACTON"
-    if isJust runacton
-      then return False
-      else do
-        hasBuildAct <- doesFileExist (path </> "Build.act")
-        hasSrcDir <- doesDirectoryExist (path </> "src")
-        return (hasBuildAct && hasSrcDir)
-
-findProjectDir :: FilePath -> IO (Maybe FilePath)
-findProjectDir path = do
-    isProjectRoot <- isActonProjectRoot path
-    if isProjectRoot
-        then return (Just path)
-        else if path == takeDirectory path  -- Check if we're at root
-            then return Nothing
-            else findProjectDir (takeDirectory path)
-
-
--- | Opaque handle for the lifetime lock of a long-running project compiler.
-data BackgroundCompilerLock = BackgroundCompilerLock FileLock FilePath
-
--- | Path to the lock held for the lifetime of a long-running project compiler.
---
--- This lock only establishes unique ownership of a background compiler
--- process such as LSP or watch mode. It does not imply that project state
--- is current, so all compile/build work must still take 'withProjectLock'.
-backgroundCompilerLockPath :: FilePath -> FilePath
-backgroundCompilerLockPath projDir =
-    joinPath [projDir, ".acton.compile.lock"]
-
--- | Acquire the long-running compiler lock for a project root, if available.
-tryBackgroundCompilerLock :: FilePath -> IO (Maybe BackgroundCompilerLock)
-tryBackgroundCompilerLock projDir = do
-    let lockPath = backgroundCompilerLockPath projDir
-    mlock <- tryLockFile lockPath Exclusive
-    return ((\lock -> BackgroundCompilerLock lock lockPath) <$> mlock)
-
--- | Release the long-running compiler lock and remove its lock file if possible.
-releaseBackgroundCompilerLock :: BackgroundCompilerLock -> IO ()
-releaseBackgroundCompilerLock (BackgroundCompilerLock lock lockPath) = do
-    unlockFile lock
-    mlock <- tryLockFile lockPath Exclusive
-    case mlock of
-      Nothing -> return ()
-      Just lock2 -> do
-        removeFile lockPath `catch` handleNotExists
-        unlockFile lock2
-  where
-    handleNotExists :: IOException -> IO ()
-    handleNotExists _ = return ()
-
--- | Path to the single per-project compiler work lock.
-projectLockPath :: FilePath -> FilePath
-projectLockPath projDir =
-    joinPath [projDir, ".acton.lock"]
-
--- | Run an action while holding the single per-project compiler work lock.
---
--- If the lock is not immediately available, run the callback once before
--- blocking until the lock can be acquired.
-withProjectLockOnWait :: FilePath -> IO () -> IO a -> IO a
-withProjectLockOnWait projDir onWait action = do
-    let lockPath = projectLockPath projDir
-    mlock <- tryLockFile lockPath Exclusive
-    case mlock of
-      Just lock -> action `finally` unlockFile lock
-      Nothing -> do
-        onWait
-        withFileLock lockPath Exclusive (\_ -> action)
-
--- | Run an action while holding the single per-project compiler work lock.
-withProjectLock :: FilePath -> IO a -> IO a
-withProjectLock projDir action =
-    withProjectLockOnWait projDir (return ()) action
-
-
--- | Compute Paths for a given source file and compile options.
--- Resolves the project root (or temp root), output dirs, and search path,
--- creating required directories along the way.
-findPaths               :: FilePath -> C.CompileOptions -> IO Paths
-findPaths actFile opts  = do execDir <- takeDirectory <$> getExecutablePath
-                             sysPath <- canonicalizePath (if null $ C.syspath opts then execDir ++ "/.." else C.syspath opts)
-                             absSrcFile <- canonicalizePath actFile
-                             (isTmp, projPath, dirInSrc) <- analyze (takeDirectory absSrcFile) []
-                             let sysTypes = joinPath [sysPath, "base", "out", "types"]
-                                 srcDir  = if isTmp then takeDirectory absSrcFile else joinPath [projPath, "src"]
-                                 projOut = joinPath [projPath, "out"]
-                                 projTypes = joinPath [projOut, "types"]
-                                 binDir  = if isTmp then srcDir else joinPath [projOut, "bin"]
-                                 modName = A.modName $ dirInSrc ++ [fileBody]
-                             -- join the search paths from command line options with the ones found in the deps directory
-                             depOverrides <- normalizeDepOverrides projPath (C.dep_overrides opts)
-                             (projName,depTypePaths) <- if isTmp then return ("",[]) else collectDepTypePaths projPath depOverrides
-                             let sPaths = [projTypes] ++ depTypePaths ++ (C.searchpath opts) ++ systemTypePaths sysPath sysTypes
-                                 modName = A.modName $ (if isTmp then dirInSrc else projName:dirInSrc) ++ [fileBody]
-                             InterfaceFiles.registerSystemTypeRoots (systemTypePaths sysPath sysTypes)
-                             createDirectoryIfMissing True binDir
-                             createDirectoryIfMissing True projOut
-                             createDirectoryIfMissing True projTypes
-                             createDirectoryIfMissing True (getModPath projTypes modName)
-                             return $ Paths sPaths sysPath sysTypes projPath projOut projTypes binDir srcDir isTmp fileExt modName projName
-  where (fileBody,fileExt) = splitExtension $ takeFileName actFile
-
-        analyze "/" ds  = do tmp <- canonicalizePath (C.tempdir opts)
-                             return (True, tmp, [])
-        analyze pre ds  = do isProjectRoot <- isActonProjectRoot pre
-                             if isProjectRoot
-                                then case ds of
-                                    [] -> return $ (False, pre, [])
-                                    "src":dirs -> return $ (False, pre, dirs)
-                                    "out":"types":dirs -> return $ (False, pre, dirs)
-                                    _ -> throwProjectError ("Source file is not in a valid project directory: " ++ joinPath ds)
-                                else analyze (takeDirectory pre) (takeFileName pre : ds)
-
--- Module helpers for multi-project builds ---------------------------------------------------------
-
--- | Derive a module name from a file path under a project's src root.
--- The result uses path segments and strips the .act extension.
-moduleNameFromFile :: FilePath -> String -> FilePath -> IO A.ModName
-moduleNameFromFile srcBase proj actFile = do
-    base <- normalizePathSafe srcBase
-    file <- normalizePathSafe actFile
-    let rel = dropExtension (makeRelative base file)
-        names = splitDirectories rel
-        mn = A.modName $ if proj `elem` special_projects then names else proj:names
-    --traceM ("## from file " ++ actFile ++ " --> " ++ prstr mn)
-    return mn
-
--- | Enumerate all .act files in a project and pair them with module names.
--- Used to seed the project module index for graph construction.
-enumerateProjectModules :: ProjCtx -> IO [(FilePath, A.ModName)]
-enumerateProjectModules ctx = do
-    exists <- doesDirectoryExist (projSrcDir ctx)
-    if not exists
-      then return []
-      else do
-        files <- getFilesRecursive (projSrcDir ctx)
-        let actFiles = filter (\f -> takeExtension f == ".act") files
-            proj = BuildSpec.specName $ projBuildSpec ctx
-            rootName = head . splitDirectories . makeRelative (projSrcDir ctx) . dropExtension
-            depNames = map fst $ projDeps ctx
-            depOverlaps = filter ((`elem` depNames) . rootName) actFiles
-
-        when (not $ null depOverlaps) $
-            throwProjectError ("Source files in project " ++ projRoot ctx ++ " overlap with declared dependencies:\n" ++
-                               concat ["    " ++ makeRelative (projRoot ctx) f ++ "\n" | f <- depOverlaps])
-
-        forM actFiles $ \f -> do
-          mn <- moduleNameFromFile (projSrcDir ctx) proj f
-          return (f, mn)
-
--- | Remove stale generated module artifacts when source modules disappear.
--- Prunes orphan .tydb/.c/.h outputs under out/types before task planning so
--- cached headers cannot mask deleted .act modules.
-pruneMissingModuleOutputs :: ProjCtx -> IO ()
-pruneMissingModuleOutputs ctx = do
-    let srcRoot = projSrcDir ctx
-        typesRoot = projTypesDir ctx
-    typesExists <- doesDirectoryExist typesRoot
-    when typesExists $ do
-      srcExists <- doesDirectoryExist srcRoot
-      srcMods <- if srcExists
-                   then do
-                     srcFiles <- getFilesRecursive srcRoot
-                     let actFiles = filter (\f -> takeExtension f == ".act") srcFiles
-                         modBases = map (normalise . dropExtension . makeRelative srcRoot) actFiles
-                     return (Data.Set.fromList modBases)
-                   else return Data.Set.empty
-      outFiles <- getFilesRecursive typesRoot
-      mapM_ (pruneFile srcMods typesRoot) outFiles
-      tyDbs <- InterfaceFiles.listInterfaceDirsRecursive typesRoot
-      mapM_ (pruneTyDb srcMods typesRoot) tyDbs
-  where
-    isRootStub rel ext =
-      ext == ".c" &&
-      (".root" `Data.List.isSuffixOf` dropExtension rel ||
-       ".test_root" `Data.List.isSuffixOf` dropExtension rel)
-
-    moduleBase rel ext
-      | isRootStub rel ext = dropExtension (dropExtension rel)
-      | otherwise = dropExtension rel
-
-    pruneFile srcMods typesRoot absFile = do
-      let ext = takeExtension absFile
-      when (ext == ".c" || ext == ".h") $ do
-        let rel = normalise (makeRelative typesRoot absFile)
-            base = dropProjDir (moduleBase rel ext)
-        unless (Data.Set.member base srcMods) $
-          removeFile absFile `catch` ignoreNotExists
-
-    pruneTyDb srcMods typesRoot absDir = do
-      let rel = normalise (makeRelative typesRoot absDir)
-          base = dropProjDir (dropExtension rel)
-      unless (Data.Set.member base srcMods) $ do
-        removePathForcibly absDir `catch` ignoreNotExists
-
-    proj = BuildSpec.specName $ projBuildSpec ctx
-
-    dropProjDir rel = case splitDirectories $ normalise (dropExtension rel) of
-                        [n] -> n
-                        n:ns | n == proj -> joinPath ns
-                        ns -> joinPath ns
-
-    ignoreNotExists :: IOException -> IO ()
-    ignoreNotExists _ = return ()
-
--- | Incremental variant: prune only modules whose changed .act file no longer exists.
-pruneMissingChangedModuleOutputs :: [ProjCtx] -> [FilePath] -> IO ()
-pruneMissingChangedModuleOutputs ctxs changedPaths = do
-    absChanged <- mapM normalizePathSafe changedPaths
-    let actPaths = filter (\p -> takeExtension p == ".act") absChanged
-    forM_ actPaths $ \actPath -> do
-      exists <- doesFileExist actPath
-      unless exists $
-        mapM_ (pruneForCtx actPath) ctxs
-  where
-    pruneForCtx actPath ctx = do
-      srcRoot <- normalizePathSafe (projSrcDir ctx)
-      let srcRoot' = addTrailingPathSeparator (normalise srcRoot)
-          actPath' = normalise actPath
-      when (Data.List.isPrefixOf srcRoot' actPath') $ do
-        let modBase = normalise (dropExtension (makeRelative srcRoot actPath'))
-            outBase = projTypesDir ctx </> modBase
-        removePathForcibly (outBase ++ InterfaceFiles.interfaceExt) `catch` ignoreNotExists
-        mapM_ (\ext -> removeFile (outBase ++ ext) `catch` ignoreNotExists) [".c", ".h"]
-
-    ignoreNotExists :: IOException -> IO ()
-    ignoreNotExists _ = return ()
-
--- | Build a search path for module interfaces for a project.
--- Includes the project's types dir, dependency types, user searchpath, and
--- the system types directory.
-searchPathForProject :: C.CompileOptions -> M.Map FilePath ProjCtx -> ProjCtx -> [FilePath]
-searchPathForProject opts projMap ctx =
-    let deps = depTypePathsFromMap projMap (projRoot ctx)
-    in [projTypesDir ctx] ++ deps ++ (C.searchpath opts) ++ systemTypePaths (projSysPath ctx) (projSysTypes ctx)
-
--- | Construct a Paths record for a module within a project context.
--- Creates output directories and ensures the types directory exists.
-pathsForModule :: C.CompileOptions -> M.Map FilePath ProjCtx -> ProjCtx -> A.ModName -> IO Paths
-pathsForModule opts projMap ctx mn = do
-    let sPaths = searchPathForProject opts projMap ctx
-        bin = joinPath [projOutDir ctx, "bin"]
-        src = projSrcDir ctx
-        proj = BuildSpec.specName $ projBuildSpec ctx
-        p = Paths sPaths (projSysPath ctx) (projSysTypes ctx) (projRoot ctx) (projOutDir ctx) (projTypesDir ctx) bin src False ".act" mn proj
-    InterfaceFiles.registerSystemTypeRoots (systemTypePaths (projSysPath ctx) (projSysTypes ctx))
-    createDirectoryIfMissing True bin
-    createDirectoryIfMissing True (projOutDir ctx)
-    createDirectoryIfMissing True (projTypesDir ctx)
-    createDirectoryIfMissing True (getModPath (projTypesDir ctx) mn)
-    return p
-
-
--- | Load a BuildSpec from Build.act.
--- Throws ProjectError on parse or validation failure.
-loadBuildSpec :: FilePath -> IO BuildSpec.BuildSpec
-loadBuildSpec dir = do
-    let actPath = joinPath [dir, "Build.act"]
-    actExists <- doesFileExist actPath
-    if actExists
-      then do
-        content <- readFile actPath
-        case BuildSpec.parseBuildActDetailed content of
-          Left err ->
-            case err of
-              BuildSpec.MissingFingerprint name -> do
-                suggestion <- suggestFingerprint name
-                throwProjectError ("Missing fingerprint in " ++ actPath ++ ".\n"
-                                   ++ "ERROR: Build.act requires `fingerprint`. For example: fingerprint = " ++ suggestion ++ "\n"
-                                   ++ "HINT: Fingerprint = CRC32(name) in the high 32 bits + random low 32 bits. You may choose a different value.")
-              BuildSpec.InvalidFingerprint name raw -> do
-                suggestion <- suggestFingerprint name
-                throwProjectError ("Invalid fingerprint " ++ raw ++ " in " ++ actPath
-                                   ++ " (project name: " ++ show name ++ ").\n"
-                                   ++ "Expected an unquoted 64-bit hex fingerprint like 0x1234abcd5678ef00.\n"
-                                   ++ "Suggested fingerprint: " ++ suggestion)
-              BuildSpec.MissingProjectName -> do
-                suggestion <- suggestProjectName dir
-                throwProjectError ("Missing project name in " ++ actPath ++ ".\n"
-                                   ++ "Add: name = " ++ show suggestion)
-              _ ->
-                throwProjectError ("Failed to parse Build.act in " ++ dir ++ ":\n"
-                                   ++ BuildSpec.renderBuildSpecParseError err)
-          Right (spec, _, _) -> validateBuildSpec actPath spec
-      else
-        throwProjectError ("Missing Build.act in " ++ dir ++ ".\n"
-                           ++ "Create Build.act with required name and fingerprint fields.")
-
-validateBuildSpec :: FilePath -> BuildSpec.BuildSpec -> IO BuildSpec.BuildSpec
-validateBuildSpec sourcePath spec = do
-    validateProjectName sourcePath (BuildSpec.specName spec)
-    validateFingerprint sourcePath (BuildSpec.specName spec) (BuildSpec.fingerprint spec)
-    return spec
-
-validateProjectName :: FilePath -> String -> IO ()
-validateProjectName sourcePath name =
-    if not (isProjectIdent name)
-      then throwProjectError ("Invalid project name '" ++ name ++ "' in " ++ sourcePath ++ ".\n"
-                              ++ "The name must be a valid Acton project name (letters, digits, underscore; cannot start with a digit).")
-      else if length name > projectNameMax
-        then throwProjectError ("Invalid project name '" ++ name ++ "' in " ++ sourcePath ++ ".\n"
-                                ++ "The name must be at most " ++ show projectNameMax ++ " characters.")
-        else return ()
-  where
-    isProjectIdent [] = False
-    isProjectIdent (c:cs) =
-      (isAlpha c || c == '_') && all isIdentChar cs
-      where
-        isIdentChar x = isAlpha x || isDigit x || x == '_'
-
-suggestProjectName :: FilePath -> IO String
-suggestProjectName dir = do
-    resolved <- normalizePathSafe dir
-    let base = takeBaseName resolved
-    if null base || base == "."
-      then return "my_project"
-      else return base
-
-suggestFingerprint :: String -> IO String
-suggestFingerprint name = do
-    let prefix = Fingerprint.fingerprintPrefixForName name
-    low <- randomRIO (1, maxBound :: Word32)
-    let fp = (fromIntegral prefix `shiftL` 32) .|. (fromIntegral low :: Word64)
-    return (Fingerprint.formatFingerprint fp)
-
-validateFingerprint :: FilePath -> String -> String -> IO ()
-validateFingerprint sourcePath name fpRaw =
-    case Fingerprint.parseFingerprint fpRaw of
-      Nothing -> do
-        suggestion <- suggestFingerprint name
-        throwProjectError ("Invalid fingerprint '" ++ fpRaw ++ "' in " ++ sourcePath
-                           ++ " (project name: " ++ show name ++ ").\n"
-                           ++ "Expected an unquoted 64-bit hex fingerprint like 0x1234abcd5678ef00.\n"
-                           ++ "Suggested fingerprint: " ++ suggestion)
-      Just fp -> do
-        let formatted = Fingerprint.formatFingerprint fp
-            expectedPrefix = Fingerprint.fingerprintPrefixForName name
-            expectedPrefixHex = Fingerprint.formatFingerprintPrefix expectedPrefix
-            actualPrefix = fromIntegral (fp `shiftR` 32)
-        if formatted == Fingerprint.fingerprintPlaceholder
-          then do
-            suggestion <- suggestFingerprint name
-            throwProjectError ("Fingerprint placeholder " ++ formatted ++ " in " ++ sourcePath
-                               ++ " is not valid.\n"
-                               ++ "Suggested fingerprint: " ++ suggestion)
-          else if expectedPrefix == actualPrefix
-            then return ()
-            else do
-              suggestion <- suggestFingerprint name
-              throwProjectError ("Fingerprint mismatch in " ++ sourcePath
-                                 ++ " for project name " ++ show name ++ ".\n"
-                                 ++ "Expected prefix: " ++ expectedPrefixHex
-                                 ++ " (CRC32 of name).\n"
-                                 ++ "Current fingerprint: " ++ formatted ++ "\n"
-                                 ++ "Renames and forks require a new fingerprint for this name.\n"
-                                 ++ "Suggested fingerprint: " ++ suggestion)
-
-
-projectNameMax :: Int
-projectNameMax = 32
-
--- | Treat drive-letter paths as absolute in addition to POSIX roots.
--- This keeps path normalization consistent on Windows hosts.
-isAbsolutePath :: FilePath -> Bool
-isAbsolutePath p =
-    isAbsolute p ||
-    case p of
-      (c:':':_) -> isAlpha c
-      _         -> False
-
--- | Normalize a path without failing if it does not exist.
--- Falls back to normalise for non-existent paths (useful for temporary builds).
-normalizePathSafe :: FilePath -> IO FilePath
-normalizePathSafe p = do
-    res <- try (canonicalizePath p) :: IO (Either IOException FilePath)
-    return $ either (const (normalise p)) id res
-
--- | Trim leading and trailing whitespace from a string.
--- Used when parsing or normalizing BuildSpec inputs.
-trim :: String -> String
-trim = f . f
-  where f = reverse . dropWhile isSpace
-
--- | Collapse "." and ".." segments without dropping leading ".." on relative paths.
--- Example: "a/b/../c/./d" becomes "a/c/d".
-collapseDots :: FilePath -> FilePath
-collapseDots p =
-    let parts = splitDirectories p
-        (root, rest) = case parts of
-                         (r:xs) | isRoot r -> (Just r, xs)
-                         xs                -> (Nothing, xs)
-        (revAcc, ups) = foldl step ([], 0) rest
-        cleaned = replicate ups ".." ++ reverse revAcc
-        prefix = maybe [] (\r -> [r]) root
-    in joinPath (prefix ++ cleaned)
-  where
-    isRoot r = r == "/" || (length r >= 2 && r !! 1 == ':')
-    step (acc, ups) comp =
-      case comp of
-        "."  -> (acc, ups)
-        ""   -> (acc, ups)
-        ".." -> case acc of
-                  (_:as) -> (as, ups)
-                  []     -> (acc, ups + 1)
-        _    -> (comp:acc, ups)
-
--- | Rebase a path against a base directory and normalize it.
--- Absolute paths are left as-is; relative paths are joined to the base.
-rebasePath :: FilePath -> FilePath -> FilePath
-rebasePath base p
-  | isAbsolutePath p = normalise p
-  | otherwise        = normalise (joinPath [base, p])
-
--- | Normalize --dep overrides relative to a base directory.
--- Absolute override paths are kept, relative paths are rebased and normalized.
-normalizeDepOverrides :: FilePath -> [(String, FilePath)] -> IO [(String, FilePath)]
-normalizeDepOverrides base overrides =
-  mapM (\(n,p) -> do
-           let absP0 = if isAbsolutePath p then p else joinPath [base, p]
-           p' <- normalizePathSafe absP0
-           return (n, p'))
-       overrides
-
--- | Apply --dep overrides (NAME=PATH) to a BuildSpec.
--- Resolves override paths relative to the given base directory when needed.
-applyDepOverrides :: FilePath -> [(String, FilePath)] -> BuildSpec.BuildSpec -> IO BuildSpec.BuildSpec
-applyDepOverrides base overrides spec = do
-    deps' <- foldM applyOne (BuildSpec.dependencies spec) overrides
-    return spec { BuildSpec.dependencies = deps' }
-  where
-    applyOne depsMap (depName, depPath) =
-      case M.lookup depName depsMap of
-        Nothing -> return depsMap
-        Just dep -> do
-          let absP0 = if isAbsolutePath depPath then depPath else joinPath [base, depPath]
-          absP <- normalizePathSafe absP0
-          validateDepOverridePath depName absP
-          let dep' = dep { BuildSpec.path = Just absP }
-          return (M.insert depName dep' depsMap)
-
-validateDepOverridePath :: String -> FilePath -> IO ()
-validateDepOverridePath depName depPath = do
-    exists <- doesDirectoryExist depPath
-    unless exists $
-      throwProjectError ("Dependency " ++ depName ++ " path does not exist: " ++ depPath ++ "\n"
-                         ++ "Hint: Local dependency paths must point to an Acton project root\n"
-                         ++ "(directory with src/ and Build.act).")
-    isProjectRoot <- isActonProjectRoot depPath
-    unless isProjectRoot $
-      throwProjectError ("Dependency " ++ depName ++ " path is not an Acton project root: " ++ depPath ++ "\n"
-                         ++ "Hint: Local dependency paths must point to an Acton project root\n"
-                         ++ "(directory with src/ and Build.act).")
-
-fetchDependencies :: C.GlobalOptions -> Paths -> [(String, FilePath)] -> IO ()
-fetchDependencies gopts paths depOverrides = do
-    if isTmp paths
-      then return ()
-      else do
-        rootSpec0 <- loadBuildSpec (projPath paths)
-        rootSpec <- applyDepOverrides (projPath paths) depOverrides rootSpec0
-        unless (C.quiet gopts) $
-          putStrLn "Resolving dependencies (fetching if missing)..."
-        home <- getHomeDirectory
-        let zigExe      = joinPath [sysPath paths, "zig", "zig"]
-            globalCache = joinPath [home, ".cache", "acton", "zig-global-cache"]
-            depsCache   = joinPath [home, ".cache", "acton", "deps"]
-            cacheDir h  = joinPath [globalCache, "p", h]
-        createDirectoryIfMissing True globalCache
-        createDirectoryIfMissing True depsCache
-        let rootPins = BuildSpec.dependencies rootSpec
-        _ <- walkProject rootPins cacheDir zigExe globalCache depsCache
-                         Data.Set.empty (projPath paths) rootSpec
-        return ()
-  where
-    walkProject rootPins cacheDir zigExe globalCache depsCache seen projDir spec = do
-      projAbs <- normalizePathSafe projDir
-      if Data.Set.member projAbs seen
-        then return seen
-        else do
-          selectedDeps <- forM (M.toList (BuildSpec.dependencies spec)) $
-            selectDependency rootPins projAbs
-          let pkgFetches = catMaybes
-                [ mkPkgFetch cacheDir zigExe globalCache name dep | (name, dep) <- selectedDeps ]
-              zigFetches = catMaybes
-                [ mkZigFetch cacheDir zigExe globalCache name dep
-                | (name, dep) <- M.toList (BuildSpec.zig_dependencies spec)
-                ]
-          results <- mapConcurrently id (pkgFetches ++ zigFetches)
-          let errs = [ e | Left e <- results ]
-          unless (null errs) $ throwProjectError (unlines errs)
-          forM_ selectedDeps $ \(name, dep) -> copyPkgDep cacheDir depsCache name dep
-          let seen' = Data.Set.insert projAbs seen
-          foldM (walkDependency rootPins cacheDir zigExe globalCache depsCache projAbs)
-                seen' selectedDeps
-
-    selectDependency rootPins base (depName, dep) = do
-      let (chosenDep, conflict) =
-            case M.lookup depName rootPins of
-              Nothing -> (dep, Nothing)
-              Just pinDep ->
-                if pinDep == dep
-                  then (dep, Nothing)
-                  else (pinDep, Just dep)
-      when (isJust conflict) $
-        unless (C.quiet gopts) $
-          putStrLn ("Warning: dependency '" ++ depName ++ "' in " ++ base
-                    ++ " overridden by root pin")
-      return (depName, chosenDep)
-
-    walkDependency rootPins cacheDir zigExe globalCache depsCache base seen (depName, dep) = do
-      depBase <- resolveDepBase base depName dep
-      depAbs <- normalizePathSafe depBase
-      depExists <- doesDirectoryExist depAbs
-      if not depExists
-        then case BuildSpec.path dep of
-               Just p | not (null p) ->
-                 throwProjectError ("Dependency " ++ depName ++ " path does not exist: " ++ depAbs ++ "\n"
-                                    ++ "Hint: Local dependency paths must point to an Acton project root\n"
-                                    ++ "(directory with src/ and Build.act).")
-               _ -> return seen
-        else do
-          depSpec0 <- loadBuildSpec depAbs
-          depSpec <- applyDepOverrides depAbs depOverrides depSpec0
-          walkProject rootPins cacheDir zigExe globalCache depsCache seen depAbs depSpec
-
-    mkPkgFetch cacheDir zigExe globalCache name dep =
-      case BuildSpec.path dep of
-        Just p | not (null p) -> Nothing
-        _ -> case (BuildSpec.url dep, BuildSpec.hash dep) of
-               (Just u, Just h) ->
-                 Just (fetchOne "pkg" name u (Just h) cacheDir zigExe globalCache)
-               (Just _, Nothing) ->
-                 Just (return (Left ("Dependency " ++ name ++ " is missing hash")))
-               _ -> Nothing
-
-    mkZigFetch cacheDir zigExe globalCache name dep =
-      case BuildSpec.zpath dep of
-        Just p | not (null p) -> Nothing
-        _ -> case (BuildSpec.zurl dep, BuildSpec.zhash dep) of
-               (Just u, Just h) ->
-                 Just (fetchOne "zig" name u (Just h) cacheDir zigExe globalCache)
-               (Just _, Nothing) ->
-                 Just (return (Left ("Zig dependency " ++ name ++ " is missing hash")))
-               _ -> Nothing
-
-    copyPkgDep cacheDir depsCache name dep =
-      case BuildSpec.path dep of
-        Just p | not (null p) -> return ()
-        _ -> case BuildSpec.hash dep of
-               Nothing -> return ()
-               Just h -> do
-                 let src = cacheDir h
-                     srcArchive = cacheArchivePath cacheDir h
-                     dst = joinPath [depsCache, name ++ "-" ++ h]
-                 exists <- doesDirectoryExist dst
-                 unless exists $ do
-                   srcOk <- doesDirectoryExist src
-                   archiveOk <- doesFileExist srcArchive
-                   unless (srcOk || archiveOk) $
-                     throwProjectError ("Dependency " ++ name ++ " not present in Zig cache after fetch: " ++ src)
-                   when (C.verbose gopts) $
-                     putStrLn ("Copying dependency " ++ name ++ " (" ++ h ++ ") from Zig cache")
-                   if srcOk
-                     then copyTree src dst
-                     else extractCachedArchive srcArchive dst
-
-    fetchOne kind name url mh cacheDir zigExe globalCache = do
-      case mh of
-        Just h -> do
-          present <- cacheEntryExists cacheDir h
-          if present
-            then do
-              unless (C.quiet gopts) $
-                putStrLn ("Using cached " ++ kind ++ " dependency " ++ name ++ " (" ++ h ++ ")")
-              return (Right h)
-            else runFetch kind name url mh cacheDir zigExe globalCache
-        Nothing ->
-          runFetch kind name url mh cacheDir zigExe globalCache
-
-    runFetch kind name url mh cacheDir zigExe globalCache = do
-      unless (C.quiet gopts) $
-        putStrLn ("Fetching " ++ kind ++ " dependency " ++ name ++ " from " ++ url)
-      if isHttpUrl url
-        then fetchViaDownloadedArchive kind name url mh cacheDir zigExe globalCache
-        else do -- other URLs, like file:// - not very common, maybe we want to constrain this somehow?
-          res <- runZigFetch zigExe globalCache url
-          case res of
-            Left ex -> return (Left ("Failed to fetch dependency " ++ name ++ ": " ++ displayException ex))
-            Right (ExitSuccess, out, _) ->
-              validateFetchOutput name mh cacheDir out
-            Right (ExitFailure _, _, err) ->
-              return (Left ("Failed to fetch dependency " ++ name ++ ":\n" ++ err))
-
-    runZigFetch :: FilePath -> FilePath -> FilePath -> IO (Either SomeException (ExitCode, String, String))
-    runZigFetch zigExe globalCache target = do
-      createDirectoryIfMissing True (globalCache </> "tmp")
-      withSystemTempDirectory "acton-zig-fetch" $ \tmp -> do
-        writeFile (tmp </> "build.zig") zigFetchBuildZig
-        let cmd = (proc zigExe ["fetch", "--global-cache-dir", globalCache, target]) { cwd = Just tmp }
-        try (readCreateProcessWithExitCode cmd "") :: IO (Either SomeException (ExitCode, String, String))
-
-    zigFetchBuildZig :: String
-    zigFetchBuildZig = unlines
-      [ "const std = @import(\"std\");"
-      , "pub fn build(b: *std.Build) void { _ = b; }"
-      ]
-
-    validateFetchOutput :: String -> Maybe String -> (String -> FilePath) -> String -> IO (Either String String)
-    validateFetchOutput name mh cacheDir out = do
-      let hashVal = trim out
-      case mh of
-        Just h | h /= hashVal ->
-          return (Left ("Hash mismatch for dependency " ++ name ++ " (expected " ++ h ++ ", got " ++ hashVal ++ ")"))
-        _ -> do
-          exists <- cacheEntryExists cacheDir hashVal
-          if exists
-            then return (Right hashVal)
-            else return (Left ("Dependency " ++ name ++ " not present in Zig cache after fetch: " ++ cacheDir hashVal))
-
-    cacheEntryExists :: (String -> FilePath) -> String -> IO Bool
-    cacheEntryExists cacheDir hashVal = do
-      dirExists <- doesDirectoryExist (cacheDir hashVal)
-      archiveExists <- doesFileExist (cacheArchivePath cacheDir hashVal)
-      return (dirExists || archiveExists)
-
-    cacheArchivePath :: (String -> FilePath) -> String -> FilePath
-    cacheArchivePath cacheDir hashVal = cacheDir hashVal ++ ".tar.gz"
-
-    extractCachedArchive :: FilePath -> FilePath -> IO ()
-    extractCachedArchive archive dst = do
-      bracketOnError (createTempDirectory (takeDirectory dst) ".acton-dep-extract")
-                     removeDirectoryRecursive $ \tmp -> do
-        let cmd = proc "tar" ["-xzf", archive, "-C", tmp, "--strip-components", "1"]
-        (code, out, err) <- readCreateProcessWithExitCode cmd ""
-        case code of
-          ExitSuccess -> renameDirectory tmp dst
-          ExitFailure _ ->
-            throwProjectError ("Failed to extract cached dependency archive " ++ archive ++ ":\n" ++ out ++ err)
-
-    fetchViaDownloadedArchive kind name depUrl mh cacheDir zigExe globalCache = do
-      dl <- downloadToLocalArchive kind name depUrl
-      case dl of
-        Left dlErr -> do
-          direct <- runZigFetch zigExe globalCache depUrl
-          case direct of
-            Left ex ->
-              return (Left ("Failed to fetch dependency " ++ name ++ ":\nDownload step failed: " ++ dlErr
-                            ++ "\nDirect zig fetch failed: " ++ displayException ex))
-            Right (ExitSuccess, out, _) ->
-              validateFetchOutput name mh cacheDir out
-            Right (ExitFailure _, _, err) ->
-              return (Left ("Failed to fetch dependency " ++ name ++ ":\nDownload step failed: " ++ dlErr
-                            ++ "\nDirect zig fetch failed:\n" ++ err))
-        Right localArchive -> do
-          fetched <- runZigFetch zigExe globalCache localArchive
-          _ <- try (removeFile localArchive) :: IO (Either IOException ())
-          case fetched of
-            Left ex ->
-              return (Left ("Failed to fetch dependency " ++ name ++ ": " ++ displayException ex))
-            Right (ExitSuccess, out, _) ->
-              validateFetchOutput name mh cacheDir out
-            Right (ExitFailure _, _, err) ->
-              return (Left ("Failed to fetch dependency " ++ name ++ ":\n" ++ err))
-
-    isHttpUrl :: String -> Bool
-    isHttpUrl depUrl =
-      let u = map toLower depUrl
-      in "http://" `isPrefixOf` u || "https://" `isPrefixOf` u
-
-    downloadToLocalArchive :: String -> String -> String -> IO (Either String FilePath)
-    downloadToLocalArchive _kind _name depUrl = do
-      parsedReq <- try (HTTP.parseRequest depUrl) :: IO (Either SomeException HTTP.Request)
-      case parsedReq of
-        Left ex -> return (Left ("Invalid dependency URL: " ++ displayException ex))
-        Right req -> do
-          let settings = HTTP.managerSetProxy (HTTP.proxyEnvironment Nothing) tlsManagerSettings
-          manager <- HTTP.newManager settings
-          opened <- try (HTTP.responseOpen req manager) :: IO (Either SomeException (HTTP.Response HTTP.BodyReader))
-          case opened of
-            Left ex -> return (Left (displayException ex))
-            Right response -> do
-              let code = statusCode (HTTP.responseStatus response)
-              if code < 200 || code >= 300
-                then do
-                  _ <- try (HTTP.responseClose response) :: IO (Either SomeException ())
-                  return (Left ("HTTP error " ++ show code))
-                else do
-                  tmpDir <- getTemporaryDirectory
-                  (tmpPath, tmpHandle) <- openBinaryTempFile tmpDir "acton-fetch"
-                  let initialSuffix = archiveSuffixForRequestResponse req response
-                  streamRes <- try (streamDownloadToFile (HTTP.responseBody response) tmpHandle) :: IO (Either SomeException BS.ByteString)
-                  _ <- try (hClose tmpHandle) :: IO (Either IOException ())
-                  _ <- try (HTTP.responseClose response) :: IO (Either SomeException ())
-                  case streamRes of
-                    Left ex -> do
-                      _ <- try (removeFile tmpPath) :: IO (Either IOException ())
-                      return (Left ("Unable to save downloaded archive: " ++ displayException ex))
-                    Right sniffBytes ->
-                      case initialSuffix <|> detectArchiveSuffixFromBytes sniffBytes of
-                        Nothing -> do
-                          _ <- try (removeFile tmpPath) :: IO (Either IOException ())
-                          return (Left "Could not determine archive type from URL path, response headers, or file bytes")
-                        Just suffix -> do
-                          let finalPath = tmpPath ++ suffix
-                          moveRes <- try (renameFile tmpPath finalPath) :: IO (Either IOException ())
-                          case moveRes of
-                            Left ex -> do
-                              _ <- try (removeFile tmpPath) :: IO (Either IOException ())
-                              return (Left ("Unable to finalize downloaded archive path: " ++ displayException ex))
-                            Right _ ->
-                              return (Right finalPath)
-
-    streamDownloadToFile :: HTTP.BodyReader -> Handle -> IO BS.ByteString
-    streamDownloadToFile bodyReader outHandle =
-      loop BS.empty
-      where
-        sniffLimit = 600
-        loop sniff = do
-          chunk <- HTTP.brRead bodyReader
-          if BS.null chunk
-            then return sniff
-            else do
-              BS.hPut outHandle chunk
-              let sniff' =
-                    if BS.length sniff >= sniffLimit
-                      then sniff
-                      else BS.take sniffLimit (sniff <> chunk)
-              loop sniff'
-
-    archiveSuffixForRequestResponse :: HTTP.Request -> HTTP.Response HTTP.BodyReader -> Maybe String
-    archiveSuffixForRequestResponse req response =
-      case archiveSuffixFromPath (B.unpack (HTTP.path req)) of
-        Just suffix -> Just suffix
-        Nothing ->
-          case archiveSuffixFromContentDisposition (lookup hContentDisposition (HTTP.responseHeaders response)) of
-            Just suffix -> Just suffix
-            Nothing -> archiveSuffixFromContentType (lookup hContentType (HTTP.responseHeaders response))
-
-    archiveSuffixFromPath :: String -> Maybe String
-    archiveSuffixFromPath rawPath =
-      let p = map toLower rawPath
-      in if ".tar.gz" `isSuffixOf` p then Just ".tar.gz"
-         else if ".tgz" `isSuffixOf` p then Just ".tgz"
-         else if ".tar.xz" `isSuffixOf` p then Just ".tar.xz"
-         else if ".txz" `isSuffixOf` p then Just ".txz"
-         else if ".tar.zst" `isSuffixOf` p then Just ".tar.zst"
-         else if ".tzst" `isSuffixOf` p then Just ".tzst"
-         else if ".tar" `isSuffixOf` p then Just ".tar"
-         else if ".zip" `isSuffixOf` p then Just ".zip"
-         else if ".jar" `isSuffixOf` p then Just ".jar"
-         else Nothing
-
-    archiveSuffixFromContentType :: Maybe B.ByteString -> Maybe String
-    archiveSuffixFromContentType mType =
-      case map toLower . trim . takeWhile (/= ';') . B.unpack <$> mType of
-        Just "application/x-tar" -> Just ".tar"
-        Just "application/gzip" -> Just ".tar.gz"
-        Just "application/x-gzip" -> Just ".tar.gz"
-        Just "application/tar+gzip" -> Just ".tar.gz"
-        Just "application/x-tar-gz" -> Just ".tar.gz"
-        Just "application/x-gtar-compressed" -> Just ".tar.gz"
-        Just "application/x-xz" -> Just ".tar.xz"
-        Just "application/zstd" -> Just ".tar.zst"
-        Just "application/zip" -> Just ".zip"
-        Just "application/x-zip-compressed" -> Just ".zip"
-        Just "application/java-archive" -> Just ".zip"
-        _ -> Nothing
-
-    archiveSuffixFromContentDisposition :: Maybe B.ByteString -> Maybe String
-    archiveSuffixFromContentDisposition mVal = do
-      headerVal <- mVal
-      filenameVal <- extractField (B.pack "filename*=") headerVal <|> extractField (B.pack "filename=") headerVal
-      archiveSuffixFromPath (B.unpack filenameVal)
-      where
-        extractField key raw =
-          let lowerRaw = B.map toLower raw
-              (prefix, rest) = B.breakSubstring key lowerRaw
-          in if B.null rest
-               then Nothing
-               else
-                 let origRest = B.drop (B.length prefix) raw
-                     val0 = B.drop (B.length key) origRest
-                     val1 = B.takeWhile (/= ';') val0
-                     val2 = stripQuotes (B.dropWhile isSpace (trimBS val1))
-                     val3 =
-                       case B.breakSubstring (B.pack "''") val2 of
-                         (_, t) | B.null t -> val2
-                         (_, t) -> B.drop 2 t
-                 in if B.null val3 then Nothing else Just val3
-        trimBS = B.dropWhileEnd isSpace . B.dropWhile isSpace
-        stripQuotes s
-          | B.length s >= 2 && B.head s == '"' && B.last s == '"' = B.tail (B.init s)
-          | otherwise = s
-
-    detectArchiveSuffixFromBytes :: BS.ByteString -> Maybe String
-    detectArchiveSuffixFromBytes bytes
-      | startsWith [0x50, 0x4b, 0x03, 0x04] bytes = Just ".zip"
-      | startsWith [0x50, 0x4b, 0x05, 0x06] bytes = Just ".zip"
-      | startsWith [0x50, 0x4b, 0x07, 0x08] bytes = Just ".zip"
-      | startsWith [0x1f, 0x8b] bytes = Just ".tar.gz"
-      | startsWith [0xfd, 0x37, 0x7a, 0x58, 0x5a, 0x00] bytes = Just ".tar.xz"
-      | startsWith [0x28, 0xb5, 0x2f, 0xfd] bytes = Just ".tar.zst"
-      | hasUstar bytes = Just ".tar"
-      | otherwise = Nothing
-      where
-        startsWith :: [Word8] -> BS.ByteString -> Bool
-        startsWith sig bs =
-          BS.length bs >= length sig && BS.take (length sig) bs == BS.pack sig
-        hasUstar bs =
-          BS.length bs >= 262 && BS.take 5 (BS.drop 257 bs) == BS.pack [0x75, 0x73, 0x74, 0x61, 0x72]
-
-    copyTree :: FilePath -> FilePath -> IO ()
-    copyTree src dst = do
-      exists <- doesDirectoryExist src
-      unless exists $ throwProjectError ("Source path for copyTree does not exist: " ++ src)
-      createDirectoryIfMissing True dst
-      entries <- listDirectory src
-      forM_ entries $ \e -> do
-        let s = src </> e
-            d = dst </> e
-        isDir <- doesDirectoryExist s
-        if isDir
-          then copyTree s d
-          else do
-            createDirectoryIfMissing True (takeDirectory d)
-            copyFile s d
-
--- | Collect dependency type directories using an existing ProjCtx map.
--- Traverses dependency edges so downstream lookups can build search paths.
-depTypePathsFromMap :: M.Map FilePath ProjCtx -> FilePath -> [FilePath]
-depTypePathsFromMap ctxs root = snd (go Data.Set.empty root)
-  where
-    go seen node =
-      case M.lookup node ctxs of
-        Nothing  -> (seen, [])
-        Just ctx ->
-          foldl' step (Data.Set.insert node seen, []) (projDeps ctx)
-
-    step (seen, acc) (_, depRoot) =
-      case M.lookup depRoot ctxs of
-        Nothing -> (Data.Set.insert depRoot seen, acc)
-        Just depCtx ->
-          if Data.Set.member depRoot seen
-            then (seen, acc)
-            else
-              let seen' = Data.Set.insert depRoot seen
-                  (seenNext, sub) = go seen' depRoot
-              in (seenNext, acc ++ [projTypesDir depCtx] ++ sub)
-
--- | Resolve a dependency's base directory from a BuildSpec PkgDep.
--- Prefers explicit path, otherwise uses the hashed cache location.
-resolveDepBase :: FilePath -> String -> BuildSpec.PkgDep -> IO FilePath
-resolveDepBase base name dep =
-    case BuildSpec.path dep of
-      Just p | not (null p) -> normalizePathSafe (rebasePath base p)
-      _ -> case BuildSpec.hash dep of
-             Just h -> do
-               home <- getHomeDirectory
-               normalizePathSafe (joinPath [home, ".cache", "acton", "deps", name ++ "-" ++ h])
-             Nothing -> throwProjectError ("Dependency " ++ name ++ " has no path or hash")
-
--- | Recursively collect out/types paths for all declared dependencies.
--- Reads Build.act and follows dependency edges.
-collectDepTypePaths :: FilePath -> [(String, FilePath)] -> IO (String,[FilePath])
-collectDepTypePaths projDir overrides = do
-  root <- normalizePathSafe projDir
-  spec <- loadBuildSpec root
-  (_, _, paths) <- go Data.Set.empty M.empty root spec
-  return (BuildSpec.specName spec, paths)
-  where
-    go seen fpMap dir spec0 = do
-      spec <- applyDepOverrides dir overrides spec0
-      let (_, fpMap1, _) = applyFingerprint dir spec fpMap
-      foldM (step dir) (seen, fpMap1, []) (M.toList (BuildSpec.dependencies spec))
-
-    step base (seen, fpMap, acc) (depName, dep) = do
-      depBase <- resolveDepBase base depName dep
-      depAbs  <- normalizePathSafe depBase
-      depExists <- doesDirectoryExist depAbs
-      if not depExists
-        then case BuildSpec.path dep of
-               Just p | not (null p) ->
-                 throwProjectError ("Dependency " ++ depName ++ " path does not exist: " ++ depAbs ++ "\n"
-                                    ++ "Hint: Local dependency paths must point to an Acton project root\n"
-                                    ++ "(directory with src/ and Build.act).")
-               _ -> return (seen, fpMap, acc)
-        else do
-          (depPath, fpMap') <- canonicalizeDep depAbs fpMap
-          let typesDir = joinPath [depPath, "out", "types"]
-          if Data.Set.member depPath seen
-            then return (seen, fpMap', acc)
-            else do
-              let seen' = Data.Set.insert depPath seen
-              spec <- loadBuildSpec depPath
-              (seenNext, fpMapNext, sub) <- go seen' fpMap' depPath spec
-              return (seenNext, fpMapNext, acc ++ [typesDir] ++ sub)
-
-    canonicalizeDep depAbs fpMap = do
-      spec0 <- loadBuildSpec depAbs
-      spec <- applyDepOverrides depAbs overrides spec0
-      let (canonPath, fpMap', _) = applyFingerprint depAbs spec fpMap
-      return (canonPath, fpMap')
 
 
 -- | Convert a module name to its source filename (path + .act).
